Renamed the input package to controls and reorganized it to share styles with other form controls.
